Explore frameworks (#72)
* refactor: Update model names in AI response functions and selection

* feat: Log cost calculation in chatcompletion_APICall and set model directly in render_data_analyst

* fix: Update model version in chatcompletion_APICall to 'gpt-4.1-mini-2025-04-14'

* fix: Update pydantic version to 2.11.3 and adjust related dependencies; refactor AI response handling and system messages

* fix: Update model names in AI response functions to 'meta/meta-llama-3.1-405b-instruct'

---------
