+++ widgets/data_analyst.py
@@ -19,8 +19,7 @@ def render_data_analyst():
         st.session_state['messages'] = [{'role': 'assistant', 'content': construct_welcome_message(), 'count': 0}]
         st.session_state['count'] = 0   
 
-    # st.session_state['model'] = st.sidebar.selectbox('Select the model:', ['meta/llama-4-scout-instruct', 'gpt-4.1-mini-2025-04-14'], index=0)
-    st.session_state['model'] = 'gpt-4.1-mini-2025-04-14'
+    st.session_state['model'] = st.sidebar.selectbox('Select the model:', ['meta/meta-llama-3.1-405b-instruct', 'gpt-4.1-mini-2025-04-14'], index=0)
     render_data_analyst_prompt_guide()
 
     with st.expander('See uploaded Datasets'):
